* Refactor how the levels dependency is loaded, to fix some localization issues * Moves reporting client to the bundles
