+++ bundle/bundle.source.js
@@ -33,4 +33,4 @@ console.info('Nightscout bundle ready');
 // Needed for Hot Module Replacement
 if(typeof(module.hot) !== 'undefined') {
     module.hot.accept()
-  }
+}
