+++ lib/client/browser-settings.js
@@ -88,7 +88,6 @@ function init (client, serverSettings, $) {
         //ignore these, they are always on for now
       } else {
         var id = 'plugin-' + plugin.name;
-        console.log('Translated plugin is',  translate(plugin.label || plugin.name));
         var dd = $('<dd><input type="checkbox" id="' + id + '" value="' + plugin.name + '"/><label for="' + id + '">' + translate(plugin.label || plugin.name) + '</label></dd>');
         showPluginsSettings.append(dd);
         dd.find('input').prop('checked', settings.showPlugins.indexOf(plugin.name) > -1);
