+++ lib/client/index.js
@@ -202,6 +202,7 @@ client.load = function load (serverSettings, callback) {
     settings: client.settings
     , extendedSettings: client.settings.extendedSettings
     , language: language
+    , levels: levels
   }).registerClientDefaults();
 
   browserSettings.loadPluginSettings(client);
@@ -221,13 +222,14 @@ client.load = function load (serverSettings, callback) {
     data: {}
     , bus: require('../bus')(client.settings, client.ctx)
     , settings: client.settings
-    , notifications: require('../notifications')(client.settings, client.ctx)
     , pluginBase: client.plugins.base(majorPills, minorPills, statusPills, bgStatus, client.tooltip, Storages.localStorage)
   };
 
   client.ctx.language = language;
   levels.translate = language.translate;
   client.ctx.levels = levels;
+
+  client.ctx.notifications = require('../notifications')(client.settings, client.ctx);
 
   client.sbx = sandbox.clientInit(client.ctx, client.now);
   client.renderer = require('./renderer')(client, d3, $);
