+++ lib/constants.json
@@ -15,5 +15,11 @@
   "ONE_HOUR": 3600000,
   "THREE_HOURS": 10800000,
   "FOUR_HOURS": 14400000,
-  "SIX_HOURS": 21600000
+  "SIX_HOURS": 21600000,
+  "LEVEL_URGENT": 2,
+  "LEVEL_WARN": 1,
+  "LEVEL_INFO": 0,
+  "LEVEL_LOW": -1,
+  "LEVEL_LOWEST": -2,
+  "LEVEL_NONE": -3
 }
