+++ lib/levels.js
@@ -1,12 +1,14 @@
 'use strict';
 
+var constants = require('./constants');
+
 var levels = {
-  URGENT: 2
-  , WARN: 1
-  , INFO: 0
-  , LOW: -1
-  , LOWEST: -2
-  , NONE: -3
+  URGENT: constants.LEVEL_URGENT
+  , WARN: constants.LEVEL_WARN
+  , INFO: constants.LEVEL_INFO
+  , LOW: constants.LEVEL_LOW
+  , LOWEST: constants.LEVEL_LOWEST
+  , NONE: constants.LEVEL_NONE
 };
 
 levels.language = require('./language')();
