+++ lib/notifications.js
@@ -1,7 +1,6 @@
 'use strict';
 
 var _ = require('lodash');
-var levels = require('./levels');
 
 var THIRTY_MINUTES = 30 * 60 * 1000;
 var DEFAULT_GROUPS = ['default'];
@@ -18,6 +17,9 @@ var Alarm = function(level, group, label) {
 var alarms = {};
 
 function init (env, ctx) {
+
+  var levels = ctx.levels;
+
   function notifications () {
     return notifications;
   }
