+++ lib/plugins/boluswizardpreview.js
@@ -1,12 +1,12 @@
 'use strict';
 
 var _ = require('lodash');
-var levels = require('../levels');
 var times = require('../times');
 
 function init (ctx) {
 
   var translate = ctx.language.translate;
+  var levels = ctx.levels;
 
   var bwp = {
     name: 'bwp'
