+++ lib/plugins/dbsize.js
@@ -1,10 +1,10 @@
 'use strict';
 
-var levels = require('../levels');
 var _ = require('lodash');
 
 function init (ctx) {
   var translate = ctx.language.translate;
+  var levels = ctx.levels;
 
   var dbsize = {
     name: 'dbsize'
