+++ lib/plugins/errorcodes.js
@@ -1,10 +1,11 @@
 'use strict';
 
 var _ = require('lodash');
-var levels = require('../levels');
 var times = require('../times');
 
-function init() {
+function init(ctx) {
+
+  var levels = ctx.levels;
 
   var errorcodes = {
     name: 'errorcodes'
