+++ lib/plugins/insulinage.js
@@ -2,10 +2,10 @@
 
 var _ = require('lodash');
 var moment = require('moment');
-var levels = require('../levels');
 
 function init(ctx) {
   var translate = ctx.language.translate;
+  var levels = ctx.levels;
 
   var iage = {
     name: 'iage'
