+++ lib/plugins/loop.js
@@ -3,13 +3,13 @@
 var _ = require('lodash');
 var moment = require('moment');
 var times = require('../times');
-var levels = require('../levels');
 
 // var ALL_STATUS_FIELDS = ['status-symbol', 'status-label', 'iob', 'freq', 'rssi']; Unused variable
 
 function init (ctx) {
   var utils = require('../utils')(ctx);
   var translate = ctx.language.translate;
+  var levels = ctx.levels;
 
   var loop = {
     name: 'loop'
