+++ lib/plugins/openaps.js
@@ -3,7 +3,6 @@
 var _ = require('lodash');
 var moment = require('moment');
 var times = require('../times');
-var levels = require('../levels');
 var consts = require('../constants');
 
 // var ALL_STATUS_FIELDS = ['status-symbol', 'status-label', 'iob', 'meal-assist', 'freq', 'rssi']; Unused variable
@@ -17,6 +16,7 @@ function init (ctx) {
   };
   var translate = ctx.language.translate;
   var firstPrefs = true;
+  var levels = ctx.levels;
 
   openaps.getClientPrefs = function getClientPrefs() {
     return ([{
