+++ lib/plugins/pump.js
@@ -2,7 +2,6 @@
 
 var _ = require('lodash');
 var moment = require('moment');
-var levels = require('../levels');
 var times = require('../times');
 
 var ALL_STATUS_FIELDS = ['reservoir', 'battery', 'clock', 'status', 'device'];
@@ -11,6 +10,7 @@ function init (ctx) {
   var translate = ctx.language.translate;
   var timeago = require('./timeago')(ctx);
   var openaps = require('./openaps')(ctx);
+  var levels = ctx.levels;
 
   var pump = {
     name: 'pump'
