+++ lib/plugins/pushover.js
@@ -3,7 +3,6 @@
 var _ = require('lodash');
 var Pushover = require('pushover-notifications');
 var request = require('request');
-var levels = require('../levels');
 var times = require('../times');
 
 function init (env) {
@@ -11,6 +10,8 @@ function init (env) {
     PRIORITY_NORMAL: 0
     , PRIORITY_EMERGENCY: 2
   };
+
+  var levels = env.levels;
 
   var pushoverAPI = setupPushover(env);
