+++ lib/plugins/sensorage.js
@@ -3,10 +3,10 @@
 var _ = require('lodash');
 var moment = require('moment');
 var times = require('../times');
-var levels = require('../levels');
 
 function init(ctx) {
   var translate = ctx.language.translate;
+  var levels = ctx.levels;
 
   var sage = {
     name: 'sage'
