+++ lib/plugins/simplealarms.js
@@ -1,15 +1,16 @@
 'use strict';
 
-var levels = require('../levels');
 var times = require('../times');
 
-function init() {
+function init(ctx) {
 
   var simplealarms = {
     name: 'simplealarms'
     , label: 'Simple Alarms'
     , pluginType: 'notification'
   };
+  
+  var levels = ctx.levels;
 
   simplealarms.checkNotifications = function checkNotifications(sbx) {
