+++ lib/plugins/timeago.js
@@ -1,12 +1,12 @@
 'use strict';
 
-var levels = require('../levels');
 var times = require('../times');
 var lastChecked = new Date();
 var lastRecoveryTimeFromSuspend = new Date("1900-01-01");
 
 function init(ctx) {
   var translate = ctx.language.translate;
+  var levels = ctx.levels;
 
   var timeago = {
     name: 'timeago',
