+++ lib/plugins/treatmentnotify.js
@@ -2,18 +2,19 @@
 
 const _ = require('lodash');
 const times = require('../times');
-const simplealarms = require('./simplealarms')();
 const crypto = require('crypto');
 
 const MANUAL_TREATMENTS = ['BG Check', 'Meal Bolus', 'Carb Correction', 'Correction Bolus'];
 
-function init() {
+function init(ctx) {
 
   const treatmentnotify = {
     name: 'treatmentnotify'
     , label: 'Treatment Notifications'
     , pluginType: 'notification'
   };
+
+  const simplealarms = require('./simplealarms')(ctx);
 
   //automated treatments from OpenAPS or Loop shouldn't trigger notifications or snooze alarms
   function filterTreatments (sbx) {
