+++ lib/plugins/upbat.js
@@ -2,10 +2,10 @@
 
 var _ = require('lodash');
 var times = require('../times');
-var levels = require('../levels');
 
 function init(ctx) {
   var translate = ctx.language.translate;
+  var levels = ctx.levels;
 
   var upbat = {
     name: 'upbat'
