+++ lib/plugins/xdripjs.js
@@ -3,9 +3,9 @@
 var _ = require('lodash');
 var moment = require('moment');
 var times = require('../times');
-var levels = require('../levels');
 
 function init(ctx) {
+  var levels = ctx.levels;
   var utils = require('../utils')(ctx);
   var firstPrefs = true;
   var lastStateNotification = null;
