+++ lib/report_plugins/index.js
@@ -2,7 +2,7 @@
 
 var _ = require('lodash');
 
-function init() {
+function init(ctx) {
   var consts = {
         SCALE_LINEAR: 0
       , SCALE_LOG: 1
@@ -44,7 +44,7 @@ function init() {
   plugins.utils = require('./utils')({
     //TODO: refactor so all this happens after init somehow
     //  until then use some defaults so things don't blow up
-    language: require('../language')()
+    language: ctx.language
     , settings: {}
   });
