+++ lib/server/bootevent.js
@@ -199,6 +199,7 @@ function boot (env, language) {
     ctx.plugins = require('../plugins')({
       settings: env.settings
       , language: ctx.language
+      , levels: ctx.levels
     }).registerServerDefaults();
 
     ctx.pushover = require('../plugins/pushover')(env);
