+++ lib/server/pushnotify.js
@@ -4,7 +4,6 @@ const _ = require('lodash');
 const crypto = require('crypto');
 const NodeCache = require('node-cache');
 
-const levels = require('../levels');
 const times = require('../times');
 
 function init (env, ctx) {
@@ -12,7 +11,7 @@ function init (env, ctx) {
   function pushnotify () {
     return pushnotify;
   }
-
+  var levels = ctx.levels;
   var receipts = new NodeCache({ stdTTL: times.hour().secs, checkperiod: times.mins(5).secs });
   var recentlySent = new NodeCache({ stdTTL: times.mins(15).secs, checkperiod: 20 });
