+++ lib/server/websocket.js
@@ -1,6 +1,5 @@
 'use strict';
 
-var levels = require('../levels');
 var times = require('../times');
 var calcData = require('../data/calcdelta');
 var ObjectID = require('mongodb').ObjectID;
@@ -10,6 +9,8 @@ function init (env, ctx, server) {
   function websocket () {
     return websocket;
   }
+
+  var levels = ctx.levels;
 
   //var log_yellow = '\x1B[33m';
   var log_green = '\x1B[32m';
