+++ lib/settings.js
@@ -1,7 +1,6 @@
 'use strict';
 
 var _ = require('lodash');
-var levels = require('./levels');
 var constants = require('./constants.json');
 
 function init () {
@@ -341,7 +340,7 @@ function init () {
   }
 
   function isUrgentHighAlarmEnabled(notify) {
-    return notify.eventName === 'high' && notify.level === levels.URGENT && settings.alarmUrgentHigh;
+    return notify.eventName === 'high' && notify.level === constants.LEVEL_URGENT && settings.alarmUrgentHigh;
   }
 
   function isHighAlarmEnabled(notify) {
@@ -349,7 +348,7 @@ function init () {
   }
 
   function isUrgentLowAlarmEnabled(notify) {
-    return notify.eventName === 'low' && notify.level === levels.URGENT && settings.alarmUrgentLow;
+    return notify.eventName === 'low' && notify.level === constants.LEVEL_URGENT && settings.alarmUrgentLow;
   }
 
   function isLowAlarmEnabled(notify) {
@@ -375,7 +374,7 @@ function init () {
       snoozeTime = settings.alarmUrgentLowMins;
     } else if (isLowAlarmEnabled(notify)) {
       snoozeTime = settings.alarmLowMins;
-    } else if (notify.level === levels.URGENT) {
+    } else if (notify.level === constants.LEVEL_URGENT) {
       snoozeTime = settings.alarmUrgentMins;
     } else {
       snoozeTime = settings.alarmWarnMins;
